Tags refactor done
