+++ src/@batch-flask/ui/tags/tag-list/tag-list.component.ts
@@ -1,4 +1,5 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component } from "@angular/core";
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnChanges } from "@angular/core";
+import { List } from "immutable";
 
 import "./tag-list.scss";
 
@@ -7,8 +8,20 @@ import "./tag-list.scss";
     templateUrl: "tag-list.html",
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class TagListComponent {
-    constructor(private changeDetector: ChangeDetectorRef) {
+export class TagListComponent implements OnChanges {
+    @Input() public tags: List<string>;
+    @Input() public maxTags: number;
+    @Input() public noTagsMessage: string;
 
+    public displayTags: List<string> = List([]);
+
+    public ngOnChanges(inputs) {
+        if (inputs.tags || inputs.maxTags) {
+            this.displayTags = List<string>(this.tags.slice(0, this.maxTags));
+        }
+    }
+
+    public trackTag(index, tag: string) {
+        return tag;
     }
 }
