+++ src/@batch-flask/ui/tags/tags.html
@@ -1,22 +1,22 @@
-<div class="main">
-    <div class="tags">
-        <bl-tag-list *ngIf="!isEditing"></bl-tag-list>
-        <form (submit)="triggerSave()" [hidden]="!isEditing">
-            <bl-form-field>
-                <input blInput #editInput placeholder="Tags(Comma separated)" [(ngModel)]="tagEditString" [ngModelOptions]="{standalone: true}">
-            </bl-form-field>
-        </form>
-    </div>
-    <div class="tags-control" *ngIf="editable">
-        <bl-clickable class="edit" matTooltip="Edit tags" (do)="edit()" *ngIf="!isEditing && !saving">
-            <i class="fa fa-pencil"></i>
-        </bl-clickable>
-        <bl-clickable class="save" matTooltip="Save" (do)="triggerSave()" *ngIf="isEditing">
-            <i class="fa fa-check"></i>
-        </bl-clickable>
-        <bl-clickable class="cancel" matTooltip="Cancel" (do)="cancel()" *ngIf="isEditing">
-            <i class="fa fa-times"></i>
-        </bl-clickable>
-        <i class="fa fa-spinner fa-spin" *ngIf="saving"></i>
-    </div>
+<div class="display" *ngIf="!isEditing && !isSaving" >
+    <bl-tag-list [tags]="tags" [maxTags]="maxTags" [noTagsMessage]="noTagsMessage"></bl-tag-list>
+    <bl-clickable class="edit" matTooltip="Edit tags" (do)="edit()" *ngIf="editable">
+        <i class="fa fa-pencil"></i>
+    </bl-clickable>
+</div>
+<div class="editor" *ngIf="isEditing || isSaving">
+    <form (submit)="triggerSave()" [hidden]="!isEditing">
+        <bl-form-field>
+            <input blInput #editInput placeholder="Tags(Comma separated)" [(ngModel)]="tagEditString" [ngModelOptions]="{standalone: true}">
+            <div blFormFieldSuffix>
+                <bl-clickable class="save" matTooltip="Save" (do)="triggerSave()" *ngIf="isEditing">
+                    <i class="fa fa-check"></i>
+                </bl-clickable>
+                <bl-clickable class="cancel" matTooltip="Cancel" (do)="cancel()" *ngIf="isEditing">
+                    <i class="fa fa-times"></i>
+                </bl-clickable>
+                <i class="fa fa-spinner fa-spin" *ngIf="saving"></i>
+            </div>
+        </bl-form-field>
+    </form>
 </div>
